+++ source/ast.hh
@@ -80,7 +80,7 @@ namespace sapc::ast {
         Location loc;
         QualifiedId name;
         std::unique_ptr<TypeRef> ref;
-        std::vector<std::unique_ptr<TypeRef>> typeParams;
+        std::vector<std::unique_ptr<TypeRef>> typeArgs;
 
         friend std::ostream& operator<<(std::ostream& os, TypeRef const& ref);
     };
@@ -123,7 +123,7 @@ namespace sapc::ast {
         std::string customTag;
         std::unique_ptr<TypeRef> baseType;
         std::vector<Field> fields;
-        std::vector<Identifier> generics;
+        std::vector<Identifier> typeParams;
         std::vector<Annotation> annotations;
     };
 
@@ -141,6 +141,7 @@ namespace sapc::ast {
         Identifier name;
         std::string customTag;
         std::vector<Field> fields;
+        std::vector<Identifier> typeParams;
         std::vector<Annotation> annotations;
     };