+++ source/compiler.cc
@@ -157,7 +157,7 @@ namespace sapc {
 
             schema::Type const* createArrayType(schema::Type const* of, Location const& loc);
             schema::Type const* createPointerType(schema::Type const* to, Location const& loc);
-            schema::Type const* createSpecializedType(schema::Type const* gen, std::vector<schema::Type const*> const& typeParams, Location const& loc);
+            schema::Type const* createSpecializedType(schema::Type const* gen, std::vector<schema::Type const*> const& typeArgs, Location const& loc);
 
             schema::Value translate(ast::Literal const& lit);
             std::unique_ptr<schema::Annotation> translate(ast::Annotation const& anno);
@@ -241,23 +241,23 @@ namespace sapc {
         type->scope = state.back().nsStack.back();
         type->location = structDecl.name.loc;
         if (structDecl.baseType != nullptr)
-            type->refType = requireType(*structDecl.baseType);
+            type->baseType = requireType(*structDecl.baseType);
         translate(type->annotations, structDecl.annotations);
 
         if (!structDecl.customTag.empty())
             applyCustomTag(*type, structDecl.customTag);
 
         // Build generics before fields, as fields might refer to a generic
-        type->generics.reserve(structDecl.generics.size());
-        for (ast::Identifier const& gen : structDecl.generics) {
-            auto* const sub = static_cast<schema::Type*>(ctx.types.emplace_back(std::make_unique<schema::Type>()).get());
-            sub->name = gen.id;
-            sub->qualifiedName = type->qualifiedName + "." + gen.id;
-            sub->kind = schema::Type::Kind::Generic;
-            sub->scope = type->scope;
-            type->generics.push_back(sub);
+        type->typeParams.reserve(structDecl.typeParams.size());
+        for (ast::Identifier const& paramDecl : structDecl.typeParams) {
+            auto* const typeParam = static_cast<schema::Type*>(ctx.types.emplace_back(std::make_unique<schema::Type>()).get());
+            typeParam->name = paramDecl.id;
+            typeParam->qualifiedName = type->qualifiedName + "." + paramDecl.id;
+            typeParam->kind = schema::Type::Kind::TypeParam;
+            typeParam->scope = type->scope;
+            type->typeParams.push_back(typeParam);
 
-            mod.types.push_back(sub);
+            mod.types.push_back(typeParam);
         }
 
         type->fields.reserve(structDecl.fields.size());
@@ -275,7 +275,7 @@ namespace sapc {
 
         auto& mod = *state.back().mod;
 
-        auto* const type = static_cast<schema::Type*>(ctx.types.emplace_back(std::make_unique<schema::Type>()).get());
+        auto* const type = static_cast<schema::TypeIndirect*>(ctx.types.emplace_back(std::make_unique<schema::TypeIndirect>()).get());
         type->name = aliasDecl.name.id;
         type->qualifiedName = qualify(type->name);
         type->kind = schema::Type::Kind::Alias;
@@ -359,6 +359,19 @@ namespace sapc {
 
         if (!unionDecl.customTag.empty())
             applyCustomTag(*type, unionDecl.customTag);
+
+        // Build generics before fields, as fields might refer to a generic
+        type->typeParams.reserve(unionDecl.typeParams.size());
+        for (ast::Identifier const& paramDecl : unionDecl.typeParams) {
+            auto* const typeParam = static_cast<schema::Type*>(ctx.types.emplace_back(std::make_unique<schema::Type>()).get());
+            typeParam->name = paramDecl.id;
+            typeParam->qualifiedName = type->qualifiedName + "." + paramDecl.id;
+            typeParam->kind = schema::Type::Kind::TypeParam;
+            typeParam->scope = type->scope;
+            type->typeParams.push_back(typeParam);
+
+            mod.types.push_back(typeParam);
+        }
 
         type->fields.reserve(unionDecl.fields.size());
         for (ast::Field const& field : unionDecl.fields)
@@ -549,14 +562,22 @@ namespace sapc {
 
         if (type.kind == schema::Type::Kind::Struct || type.kind == schema::Type::Kind::Attribute || type.kind == schema::Type::Kind::Union) {
             auto const& typeAggr = static_cast<schema::TypeAggregate const&>(type);
-            makeAvailable(typeAggr.refType);
+
+            makeAvailable(typeAggr.baseType);
+
             for (auto const& field : typeAggr.fields)
                 makeAvailableRecurse(*field);
+
+            for (auto const* typeParam : typeAggr.typeParams)
+                makeAvailableRecurse(*typeParam);
         }
         else if (type.kind == schema::Type::Kind::Alias || type.kind == schema::Type::Kind::Pointer || type.kind == schema::Type::Kind::Array || type.kind == schema::Type::Kind::Specialized) {
-            makeAvailable(type.refType);
-            for (auto const* generic : type.generics)
-                makeAvailableRecurse(*generic);
+            auto const& typeInd = static_cast<schema::TypeIndirect const&>(type);
+
+            makeAvailable(typeInd.refType);
+
+            for (auto const* typeArg : typeInd.typeArgs)
+                makeAvailableRecurse(*typeArg);
         }
     }
 
@@ -600,15 +621,15 @@ namespace sapc {
             return nullptr;
         case ast::TypeRef::Kind::Generic:
             if (auto const* type = resolveType(*ref.ref, scope); type != nullptr) {
-                std::vector<schema::Type const*> typeParams;
-                typeParams.reserve(ref.typeParams.size());
-                for (auto const& refParam : ref.typeParams) {
-                    auto const* typeParam = requireType(*refParam, scope);
-                    if (typeParam == nullptr)
+                std::vector<schema::Type const*> typeArgs;
+                typeArgs.reserve(ref.typeArgs.size());
+                for (auto const& refTypeArg : ref.typeArgs) {
+                    auto const* typeArg = requireType(*refTypeArg, scope);
+                    if (typeArg == nullptr)
                         return nullptr;
-                    typeParams.push_back(typeParam);
+                    typeArgs.push_back(typeArg);
                 }
-                return createSpecializedType(type, typeParams, ref.loc);
+                return createSpecializedType(type, typeArgs, ref.loc);
             }
             return nullptr;
         default:
@@ -632,12 +653,12 @@ namespace sapc {
         }
         else if (scope->kind == schema::Type::Kind::Struct) {
             auto const& typeAggr = *static_cast<schema::TypeAggregate const*>(scope);
-            if (scope->refType != nullptr)
-                if (auto const rs = findLocal(qualId, scope->refType); rs.kind != Resolve::Kind::Empty)
+            if (typeAggr.baseType != nullptr)
+                if (auto const rs = findLocal(qualId, typeAggr.baseType); rs.kind != Resolve::Kind::Empty)
                     return rs;
-            for (auto const* gen : typeAggr.generics)
-                if (gen->name == qualId.front().id)
-                    return Resolve{ gen };
+            for (auto const* typeParam : typeAggr.typeParams)
+                if (typeParam->name == qualId.front().id)
+                    return Resolve{ typeParam };
         }
 
         return {};
@@ -772,7 +793,7 @@ namespace sapc {
 
         auto& top = state.back();
 
-        auto* arr = static_cast<schema::Type*>(ctx.types.emplace_back(std::make_unique<schema::Type>()).get());
+        auto* arr = static_cast<schema::TypeIndirect*>(ctx.types.emplace_back(std::make_unique<schema::TypeIndirect>()).get());
         top.mod->types.push_back(arr);
 
         arr->name = of->name;
@@ -798,7 +819,7 @@ namespace sapc {
 
         auto& top = state.back();
 
-        auto* ptr = static_cast<schema::Type*>(ctx.types.emplace_back(std::make_unique<schema::Type>()).get());
+        auto* ptr = static_cast<schema::TypeIndirect*>(ctx.types.emplace_back(std::make_unique<schema::TypeIndirect>()).get());
         top.mod->types.push_back(ptr);
 
         ptr->name = to->name;
@@ -813,12 +834,12 @@ namespace sapc {
         return pointerTypeMap.emplace(to, ptr).first->second;
     }
 
-    schema::Type const* Compiler::createSpecializedType(schema::Type const* gen, std::vector<schema::Type const*> const& typeParams, Location const& loc) {
+    schema::Type const* Compiler::createSpecializedType(schema::Type const* gen, std::vector<schema::Type const*> const& typeArgs, Location const& loc) {
         assert(gen != nullptr);
 
         size_t specHash = (std::hash<std::string>{}(gen->qualifiedName));
-        for (auto const* param : typeParams)
-            specHash = hash_combine(param->qualifiedName, specHash);
+        for (auto const* typeArg : typeArgs)
+            specHash = hash_combine(typeArg->qualifiedName, specHash);
 
         {
             auto const it = specializedTypeMap.find(specHash);
@@ -828,12 +849,12 @@ namespace sapc {
 
         auto& top = state.back();
 
-        auto* spec = static_cast<schema::Type*>(ctx.types.emplace_back(std::make_unique<schema::Type>()).get());
+        auto* spec = static_cast<schema::TypeIndirect*>(ctx.types.emplace_back(std::make_unique<schema::TypeIndirect>()).get());
         top.mod->types.push_back(spec);
 
         std::string genName = "<";
-        for (auto const* param : typeParams)
-            genName += param->qualifiedName;
+        for (auto const* typeArg : typeArgs)
+            genName += typeArg->qualifiedName;
         genName += '>';
 
         spec->name = gen->name;
@@ -845,9 +866,9 @@ namespace sapc {
         spec->scope = gen->scope;
         spec->location = loc;
 
-        spec->generics.reserve(typeParams.size());
-        for (auto const* param : typeParams)
-            spec->generics.push_back(param);
+        spec->typeArgs.reserve(typeArgs.size());
+        for (auto const* typeArg : typeArgs)
+            spec->typeArgs.push_back(typeArg);
 
         return specializedTypeMap.emplace(specHash, spec).first->second;
     }