+++ source/schema.hh
@@ -63,7 +63,7 @@ namespace sapc::schema {
         enum Kind {
             Simple,
             Struct,
-            Generic,
+            TypeParam,
             Specialized,
             Union,
             Attribute,
@@ -79,16 +79,21 @@ namespace sapc::schema {
         std::string qualifiedName;
         Location location;
         Namespace const* scope = nullptr;
-        Type const* refType = nullptr; // arrays, pointers, aliases, specialized, base for aggregate
-        std::vector<Type const*> generics; // placeholders for generic types; type params for specialized types
     };
 
     struct TypeAggregate : Type {
+        Type const* baseType = nullptr;
         std::vector<std::unique_ptr<Field>> fields;
+        std::vector<Type const*> typeParams;
     };
 
     struct TypeEnum : Type {
         std::vector<std::unique_ptr<EnumItem>> items;
+    };
+
+    struct TypeIndirect : Type {
+        Type const* refType = nullptr; // arrays, pointers, aliases, specialized
+        std::vector<Type const*> typeArgs; // types arguments for specialized types (generics)
     };
 
     struct Constant : Annotated {