+++ NEWS.md
@@ -7,7 +7,8 @@ Version 0.15-beta (WIP)
  - Importing modules includes their custom tags in grammar parsing
  - Fix enums to apply custom tags
  - Custom tags for unions, aliases, and constants
- - Rename $sapc, $sapc.customtag, and $sapc.typeid to $core, $customtag, $typeid, respectively
+ - Renamed $sapc, $sapc.customtag, and $sapc.typeid to $core, $customtag, $typeid, respectively
+ - Renamed 'generic' type kind to 'typeparam', 'generics' array to 'typeParams', and old 'typeParams' to 'typeArgs'
 
 Version 0.14-beta
 -----------------